.circleci/deploy_ecr/requirements.txt
aws-cdk-lib==2.29.1
constructs>=10.0.0,<11.0.0
ub-ecr-cdk==1.1.221
